# Optimize spatial filters