Significantly improved performance of spatial filters. Added Erode and Dilate tests to ensure no change in output.

```
benchmark              old ns/op      new ns/op      delta
BenchmarkMedian1-8     8960811        6144771        -31.43%
BenchmarkMedian4-8     371908494      222312970      -40.22%
BenchmarkMedian8-8     4289995847     2485732320     -42.06%

benchmark              old allocs     new allocs     delta
BenchmarkMedian1-8     65542          65544          +0.00%
BenchmarkMedian4-8     65542          65544          +0.00%
BenchmarkMedian8-8     65545          65546          +0.00%

benchmark              old bytes     new bytes     delta
BenchmarkMedian1-8     3670588       3678857       +0.23%
BenchmarkMedian4-8     23606308      23626594      +0.09%
BenchmarkMedian8-8     84542640      84539120      -0.00%
```